feat(WIP):取引先
Restrict partner deletion based on category association

Added a check in `PartnerService` to prevent the deletion of partners that are associated with a partner category. Introduced a new `BusinessException` to handle these validations gracefully. Updated tests to cover various deletion scenarios, including restricted cases.

+++ app/backend/api/src/main/java/com/example/sms/service/master/partner/PartnerService.java
@@ -1,7 +1,9 @@
 package com.example.sms.service.master.partner;
 
 import com.example.sms.domain.model.master.partner.Partner;
+import com.example.sms.domain.model.master.partner.PartnerCategoryList;
 import com.example.sms.domain.model.master.partner.PartnerList;
+import com.example.sms.service.BusinessException;
 import com.github.pagehelper.PageInfo;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,9 +15,11 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional
 public class PartnerService {
     final PartnerRepository partnerRepository;
+    final PartnerCategoryRepository partnerCategoryRepository;
 
-    public PartnerService(PartnerRepository partnerRepository) {
+    public PartnerService(PartnerRepository partnerRepository, PartnerCategoryRepository partnerCategoryRepository) {
         this.partnerRepository = partnerRepository;
+        this.partnerCategoryRepository = partnerCategoryRepository;
     }
 
     /**
@@ -42,8 +46,14 @@ public class PartnerService {
     /**
      * 取引先削除
      */
-    public void delete(Partner partner) {
-        partnerRepository.deleteById(partner);
+    public void delete(Partner partner) throws BusinessException {
+        PartnerCategoryCriteria criteria = PartnerCategoryCriteria.builder().partnerCode(partner.getPartnerCode().getValue()).build();
+        PartnerCategoryList result = partnerCategoryRepository.search(criteria);
+        if (result.size() > 0) {
+            throw new BusinessException("取引先分類が登録されているため削除できません。");
+        } else {
+            partnerRepository.deleteById(partner);
+        }
     }
 
     /**
